scheduler/requirements.txt
appnope==0.1.0
APScheduler==3.6.3
asgiref==3.3.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1.3
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.3.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.25.11
wcwidth==0.2.5
zipp==3.4.0
